initial thoughts on new api
